model/EfficientNet/requirements.txt
flask
flask-cors
torch
torchvision
timm
pillow
requests
